WIP: extract Render classes to top level, remove now unused renderStats method
