+++ src/test/java/org/openjdk/shenandoah/RenderLegendTest.java
@@ -39,7 +39,7 @@ public class RenderLegendTest {
         Graphics2D g = img.createGraphics();
         g.setColor(Color.WHITE);
         g.fillRect(0, 0, 300, 700);
-        new ShenandoahVisualizer.RenderPlayback(new DataLogProvider("src/test/resources/regions-6425.log", new SnapshotHistory()), null, new ToolbarPanel(true)).renderLegend(g);
+        new RenderPlayback(new DataLogProvider("src/test/resources/regions-6425.log", new SnapshotHistory()), null, new ToolbarPanel(true)).renderLegend(g);
         ImageIO.write(img, "png", new File("legend.png"));
     }
